+++ x/tally/keeper/filter_test.go
@@ -643,7 +643,7 @@ func TestFilter(t *testing.T) {
 				sort.Strings(tt.reveals[i].ProxyPubKeys)
 			}
 
-			result, err := f.tallyKeeper.ApplyFilter(f.Context(), filter, tt.reveals, int64(len(tt.reveals)))
+			result, err := f.tallyKeeper.ApplyFilter(f.Context(), filter, tt.reveals, uint16(len(tt.reveals)))
 			require.ErrorIs(t, err, tt.wantErr)
 			if tt.consPubKeys == nil {
 				require.Nil(t, nil, result.ProxyPubKeys)
